Added a catch for unreliable requests returned from github.
