+++ leaderboard-frontend/src/components/Leaderboard/LeaderBoard.js
@@ -7,6 +7,8 @@ import WeeklyLeaderboard from '../LeaderboardComponents/WeeeklyLeaderboard';
 import WeeklyData from '../LeaderboardComponents/WeeklyData'
 import OverallLeaderboard from '../LeaderboardComponents/OverallLeaderboard';
 import OverallData from "../LeaderboardComponents/OverallData";
+import {connectAsync} from 'iguazu'
+import {queryGithub, queryStudents} from "../../actions";
 
 class LeaderBoard extends Component {
     constructor(props) {
@@ -23,6 +25,57 @@ class LeaderBoard extends Component {
          }
     }
     render() {
+        if (localStorage.getItem("invalid")) {
+            localStorage.removeItem("invalid");
+            this.props.props.props.history.push('/')
+        }
+
+        if (this.props.isLoading()) {
+            if (localStorage.getItem("invalid")) {
+                localStorage.removeItem("token");
+                localStorage.removeItem("invalid");
+                this.props.props.props.history.push('/')
+            }
+            return <div>Loading...</div>
+        }
+
+        if (this.props.loadedWithErrors()) {
+            return <div>Oh no! Something went wrong</div>
+        }
+        let count = []
+        const studentsObject = [];
+        this.props.students.forEach((each, i) => {
+            if (each.classname === this.props.match.params.classname) {
+                studentsObject.push({
+                    each
+                })
+            }
+
+        })
+        const gitObject = [];
+        this.props.data.gitData.forEach((git, x) => {
+            studentsObject.forEach(student => {
+                //GitHub sometimes returns a null value
+                //This is due to some requests having a valid handle
+                //but github fails to return the users data
+                //the .then response is a null object
+                //because it's a null object, git.FullName will crash the application
+                //the quick solution is (git !== null)
+                if (git !== null) {
+                    if (git.FullName === student.each.firstname + ' ' + student.each.lastname) {
+                        gitObject.push(git)
+                    }
+                }
+
+            })
+        })
+        if (gitObject.length === 0) {
+            return (
+                <div>
+                    <h1>No Students</h1>
+                </div>
+            )
+        }
         return (
             <div className="App">
               <p></p>
@@ -31,17 +84,29 @@ class LeaderBoard extends Component {
               </div>
               <div>
                 {/*<WeeklyLeaderboard />*/}
-                <WeeklyData props={this.props} />
+                <WeeklyData props={this.props} gitObject={gitObject} data={this.props.data}  students={this.props.students} />
               </div>
               <div class="ui horizontal divider"></div>
               <div>
                 {/*<OverallLeaderboard />*/}
-                <OverallData props={this.props}/>
+                <OverallData props={this.props} gitObject={gitObject} data={this.props.data} students={this.props.students}/>
               </div>
 
             </div>
           );
     }
 }
+export function loadDataAsProps({store, ownProps}) {
+    const {dispatch} = store;
 
-export default LeaderBoard;
+    const path = "/"; // Use the actual path when it's created as needed
+    console.log(ownProps);
+    return {
+        // classdata: () => dispatch(queryAllMyData(path)),
+        students: () => dispatch(queryStudents()),
+        data: () => dispatch(queryGithub())
+    };
+}
+
+
+export default connectAsync({loadDataAsProps})(LeaderBoard);
